third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/template.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] shadowRootCustomElements reflects as string
  assert_equals: expected (string) "" but got (undefined) undefined
[FAIL] Serializing a ShadowRoot with a null registry
  assert_equals: expected "<div><template shadowrootmode=\\"open\\" shadowrootserializable=\\"\\" shadowrootcustomelements=\\"\\"></template></div>" but got "<div><template shadowrootmode=\\"open\\" shadowrootserializable=\\"\\"></template></div>"
[FAIL] Serializing a ShadowRoot with a registry that differs from its host
  registry.initialize is not a function
Harness: the test ran to completion.
